Windows compat tasks (#20)
* make binny taskfile parseable on Windows

When running tests on Windows, core utils like grep may not be present,
or may not behave the same way. Therefore, implement inline shell logic
in python and invoke it from the task file.

Signed-off-by: Will Murphy <will.murphy@anchore.com>

* enable unit tests to run on Windows

Signed-off-by: Will Murphy <will.murphy@anchore.com>

* skip or fix some unit tests on Windows

Signed-off-by: Will Murphy <will.murphy@anchore.com>

* chore: windows compatibility due to open file failures

Signed-off-by: Keith Zantow <kzantow@gmail.com>

* Fix and skip more unit tests

For now, simply do not support hosted shell script installer on Windows.

Signed-off-by: Will Murphy <will.murphy@anchore.com>

* prevent CRLF normalization of test fixtures

There are text files under tool/testdata/store that are used as binary
fixtures, including taking SHA256 of the file during unit tests. Prevent
git from applying CRLF autocorrection to these files by removing the
"text" attribute from them so that their SHA256 will be stable between
Windows and other operating systems.

Signed-off-by: Will Murphy <will.murphy@anchore.com>

* Enable Windows unit tests

This is a stepping stone towards actual Windows support. Eventually, all
validations should be run, and release artifacts should be built for
Windows.

Signed-off-by: Will Murphy <will.murphy@anchore.com>

* Fix workflow actions yaml syntax

Signed-off-by: Will Murphy <will.murphy@anchore.com>

* set Windows-specific coverage threshold

There are a few paths that we know are not supported on Windows, so a
lower coverage threshold is acceptable.

Signed-off-by: Will Murphy <will.murphy@anchore.com>

---------

Signed-off-by: Will Murphy <will.murphy@anchore.com>
Signed-off-by: Keith Zantow <kzantow@gmail.com>
Co-authored-by: Keith Zantow <kzantow@gmail.com>